Add indicator to show current indicator using `libxkbcommon`
The indicator is based on XCB-XKB and `libxkbcommon`. It can be tested
using the following configuration:

```
bar-list = ["bar"];
bar: {
  position: "top";
  block-list: ["ya_xkb"];
  ya_xkb: {
    exec: "YABAR_KEYBOARD_LAYOUT";
    interval: 1;
  };
}
```

The followinng new dependencies have been introduced:

- libxkbcommon-dev
- libxkbcommon-x11-dev
- libxcb-xkb-dev

Both work with libxcb and allow the usage of XKB through an XCB session.
In this case the session `ya.c` created by `src/ya_exec.c` can be used
for this indicator, so no extra X11 logic is needed.